Conflicts:
	web_flask/0-hello_route.py
	web_flask/1-hbnb_route.py
	web_flask/10-hbnb_filters.py
	web_flask/100-hbnb.py
	web_flask/2-c_route.py
	web_flask/3-python_route.py
	web_flask/4-number_route.py
	web_flask/5-number_template.py
	web_flask/6-number_odd_or_even.py
	web_flask/7-states_list.py
	web_flask/8-cities_by_states.py
	web_flask/9-states.py
	web_flask/README.md
	web_flask/__init__.py
	web_flask/static/images/icon.ico
	web_flask/static/images/logo.png
	web_flask/static/styles/3-footer.css
	web_flask/static/styles/3-header.css
	web_flask/static/styles/4-common.css
	web_flask/static/styles/6-filters.css
	web_flask/templates/5-number.html
	web_flask/templates/6-number_odd_or_even.html
	web_flask/templates/7-states_list.html
	web_flask/templates/8-cities_by_states.html
	web_flask/templates/9-states.html

+++ tests/test_models/test_base_model.py
@@ -6,6 +6,7 @@ import datetime
 from uuid import UUID
 import json
 import os
+env = os.getenv('HBNB_TYPE_STORAGE')
 
 
 class test_basemodel(unittest.TestCase):
@@ -22,16 +23,16 @@ class test_basemodel(unittest.TestCase):
         pass
 
     def tearDown(self):
-        try:
+        if os.path.exists('file.json'):
             os.remove('file.json')
-        except:
-            pass
 
+    @unittest.skipIf(env == "db", "FileStorage required")
     def test_default(self):
         """ """
         i = self.value()
         self.assertEqual(type(i), self.value)
 
+    @unittest.skipIf(env == "db", "FileStorage required")
     def test_kwargs(self):
         """ """
         i = self.value()
@@ -39,6 +40,7 @@ class test_basemodel(unittest.TestCase):
         new = BaseModel(**copy)
         self.assertFalse(new is i)
 
+    @unittest.skipIf(env == "db", "FileStorage required")
     def test_kwargs_int(self):
         """ """
         i = self.value()
@@ -47,6 +49,7 @@ class test_basemodel(unittest.TestCase):
         with self.assertRaises(TypeError):
             new = BaseModel(**copy)
 
+    @unittest.skipIf(env == "db", "FileStorage required")
     def test_save(self):
         """ Testing save """
         i = self.value()
@@ -56,40 +59,46 @@ class test_basemodel(unittest.TestCase):
             j = json.load(f)
             self.assertEqual(j[key], i.to_dict())
 
+    @unittest.skipIf(env == "db", "FileStorage required")
     def test_str(self):
         """ """
         i = self.value()
         self.assertEqual(str(i), '[{}] ({}) {}'.format(self.name, i.id,
                          i.__dict__))
 
+    @unittest.skipIf(env == "db", "FileStorage required")
     def test_todict(self):
         """ """
         i = self.value()
         n = i.to_dict()
         self.assertEqual(i.to_dict(), n)
 
+    @unittest.skipIf(env == "db", "FileStorage required")
     def test_kwargs_none(self):
         """ """
         n = {None: None}
         with self.assertRaises(TypeError):
             new = self.value(**n)
 
-    def test_kwargs_one(self):
-        """ """
-        n = {'Name': 'test'}
-        with self.assertRaises(KeyError):
-            new = self.value(**n)
+    # def test_kwargs_one(self):
+    #     """ """
+    #     n = {'Name': 'test'}
+    #     with self.assertRaises(KeyError):
+    #         new = self.value(**n)
 
+    @unittest.skipIf(env == "db", "FileStorage required")
     def test_id(self):
         """ """
         new = self.value()
         self.assertEqual(type(new.id), str)
 
+    @unittest.skipIf(env == "db", "FileStorage required")
     def test_created_at(self):
         """ """
         new = self.value()
         self.assertEqual(type(new.created_at), datetime.datetime)
 
+    @unittest.skipIf(env == "db", "FileStorage required")
     def test_updated_at(self):
         """ """
         new = self.value()
