+++ tests/test_models/test_engine/test_file_storage.py
@@ -6,6 +6,8 @@ from models import storage
 import os
 
 
+@unittest.skipIf(os.getenv('HBNB_TYPE_STORAGE') == "db",
+                 "FileStorage required")
 class test_fileStorage(unittest.TestCase):
     """ Class to test the file storage method """
 
@@ -19,10 +21,9 @@ class test_fileStorage(unittest.TestCase):
 
     def tearDown(self):
         """ Remove storage file at end of tests """
-        try:
+
+        if os.path.exists('file.json'):
             os.remove('file.json')
-        except:
-            pass
 
     def test_obj_list_empty(self):
         """ __objects is initially empty """
@@ -31,6 +32,7 @@ class test_fileStorage(unittest.TestCase):
     def test_new(self):
         """ New object is correctly added to __objects """
         new = BaseModel()
+        new.save()
         for obj in storage.all().values():
             temp = obj
         self.assertTrue(temp is obj)
@@ -63,6 +65,7 @@ class test_fileStorage(unittest.TestCase):
     def test_reload(self):
         """ Storage file is successfully loaded to __objects """
         new = BaseModel()
+        new.save()
         storage.save()
         storage.reload()
         for obj in storage.all().values():
@@ -97,6 +100,7 @@ class test_fileStorage(unittest.TestCase):
     def test_key_format(self):
         """ Key is properly formatted """
         new = BaseModel()
+        new.save()
         _id = new.to_dict()['id']
         for key in storage.all().keys():
             temp = key
@@ -105,5 +109,4 @@ class test_fileStorage(unittest.TestCase):
     def test_storage_var_created(self):
         """ FileStorage object storage created """
         from models.engine.file_storage import FileStorage
-        print(type(storage))
         self.assertEqual(type(storage), FileStorage)
